Remove unused hk2017 test helper modules
