+++ crates/library/src/algorithms/hk2017/tests/mod.rs
@@ -1,0 +1,6 @@
+mod determinism;
+mod finite_difference_checks;
+mod literature_checks;
+mod monotonicity_conformality;
+mod properties;
+mod pruning_equivalence;
